requirements.txt
discord.py
openai
anthropic
redis
python-dotenv